requirements.txt
accelerate==0.21.0
aiofiles==23.2.1
aiohttp==3.9.5
aiosignal==1.3.1
altair==5.3.0
annotated-types==0.6.0
antlr4-python3-runtime==4.9.3
anyio==4.3.0
asttokens==2.4.1
attrs==23.2.0
av==11.0.0
bitsandbytes==0.41.0
braceexpand==0.1.7
certifi==2022.12.7
charset-normalizer==2.1.1
click==8.1.7
cmake==3.29.2
comm==0.2.1
contourpy==1.2.0
cycler==0.12.1
datasets==2.16.1
debugpy==1.8.0
decorator==4.4.2
decord==0.6.0
dill==0.3.7
distro==1.9.0
easydict==1.11
einops==0.6.1
einops-exts==0.0.4
et-xmlfile==1.1.0
executing==2.0.1
fastapi==0.110.2
ffmpy==0.3.2
filelock==3.9.0
fonttools==4.47.0
frozenlist==1.4.1
fsspec==2023.10.0
fvcore==0.1.5.post20221221
gradio==4.16.0
gradio_client==0.8.1
h11==0.14.0
httpcore==0.17.3
httpx==0.24.0
huggingface-hub==0.20.1
hydra-core==1.3.2
idna==3.4
imageio==2.33.1
imageio-ffmpeg==0.4.9
importlib_resources==6.4.0
iopath==0.1.10
ipykernel==6.28.0
ipython==8.20.0
jedi==0.19.1
Jinja2==3.1.2
joblib==1.3.2
jsonschema==4.21.1
jsonschema-specifications==2023.12.1
jupyter_client==8.6.0
jupyter_core==5.7.1
kiwisolver==1.4.5
linkify-it-py==2.0.3
lit==18.1.3
-e git+https://github.com/haotian-liu/LLaVA.git@3e337ad269da3245643a2724a1d694b5839c37f9#egg=llava
lmdb==1.4.1
markdown-it-py==2.2.0
markdown2==2.4.13
MarkupSafe==2.1.3
matplotlib==3.8.2
matplotlib-inline==0.1.6
mdit-py-plugins==0.3.3
mdurl==0.1.2
moviepy==1.0.3
mpmath==1.3.0
multidict==6.0.5
multiprocess==0.70.15
nest-asyncio==1.5.8
networkx==3.0
numpy==1.24.1
nvidia-cublas-cu11==11.10.3.66
nvidia-cublas-cu12==12.1.3.1
nvidia-cuda-cupti-cu11==11.7.101
nvidia-cuda-cupti-cu12==12.1.105
nvidia-cuda-nvrtc-cu11==11.7.99
nvidia-cuda-nvrtc-cu12==12.1.105
nvidia-cuda-runtime-cu11==11.7.99
nvidia-cuda-runtime-cu12==12.1.105
nvidia-cudnn-cu11==8.5.0.96
nvidia-cudnn-cu12==8.9.2.26
nvidia-cufft-cu11==10.9.0.58
nvidia-cufft-cu12==11.0.2.54
nvidia-curand-cu11==10.2.10.91
nvidia-curand-cu12==10.3.2.106
nvidia-cusolver-cu11==11.4.0.1
nvidia-cusolver-cu12==11.4.5.107
nvidia-cusparse-cu11==11.7.4.91
nvidia-cusparse-cu12==12.1.0.106
nvidia-nccl-cu11==2.14.3
nvidia-nccl-cu12==2.18.1
nvidia-nvjitlink-cu12==12.4.127
nvidia-nvtx-cu11==11.7.91
nvidia-nvtx-cu12==12.1.105
omegaconf==2.3.0
openai==1.8.0
opencv-python==4.8.1.78
openpyxl==3.1.2
orjson==3.10.1
packaging==23.2
pandas==2.1.4
parameterized==0.9.0
parso==0.8.3
peft==0.5.0
pexpect==4.9.0
Pillow==9.3.0
platformdirs==4.1.0
portalocker==2.8.2
proglog==0.1.10
prompt-toolkit==3.0.43
protobuf==4.25.1
psutil==5.9.7
ptyprocess==0.7.0
pure-eval==0.2.2
pyarrow==16.0.0
pyarrow-hotfix==0.6
pydantic==2.7.1
pydantic_core==2.18.2
pydub==0.25.1
Pygments==2.17.2
pyparsing==3.1.1
python-dateutil==2.8.2
python-multipart==0.0.9
pytorchvideo==0.1.5
pytz==2023.3.post1
PyYAML==6.0.1
pyzmq==25.1.2
referencing==0.34.0
regex==2023.12.25
requests==2.28.1
rich==13.7.1
rpds-py==0.18.0
ruff==0.4.3
safetensors==0.4.1
scikit-learn==1.2.2
scipy==1.11.4
seaborn==0.13.1
semantic-version==2.10.0
sentencepiece==0.1.99
shellingham==1.5.4
shortuuid==1.0.13
six==1.16.0
sniffio==1.3.1
stack-data==0.6.3
starlette==0.37.2
svgwrite==1.4.3
sympy==1.12
tabulate==0.9.0
tensorboardX==2.6.2.2
termcolor==2.4.0
threadpoolctl==3.2.0
timm==0.6.13
tokenizers==0.15.1
tomlkit==0.12.0
toolz==0.12.1
torch==2.1.2
torchaudio==2.1.2+cu118
torchvision==0.16.2
tornado==6.4
tqdm==4.66.1
traitlets==5.14.1
transformers==4.37.2
triton==2.1.0
typer==0.12.3
typing_extensions==4.11.0
tzdata==2023.4
uc-micro-py==1.0.3
ujson==5.9.0
urllib3==1.26.13
uvicorn==0.29.0
-e git+https://github.com/Efficient-Large-Model/VILA.git@eaadb1e55a088978ce06abb6242edc251fb4665a#egg=vila
wavedrom==2.0.3.post3
wcwidth==0.2.13
webdataset==0.2.86
websockets==11.0.3
xxhash==3.4.1
yacs==0.1.8
yarl==1.9.4
